celery_worker/requirements.txt
acres==0.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
billiard==4.2.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
dnspython==2.7.0
easyocr==1.7.2
etelemetry==0.3.1
filelock==3.18.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.28.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
kombu==5.5.4
lazy_loader==0.4
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nibabel==5.3.2
ninja==1.11.1.4
nipype==1.10.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
pillow==11.3.0
prompt_toolkit==3.0.51
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
pymongo==4.13.2
PyMuPDF==1.26.3
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
redis==6.2.0
requests==2.32.4
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==8.5.0
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
